+++ libs/sdk-core/CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add `question_answers` property to `ExtractedData` interface
   - Add `QUESTION_ANSWERS` type to `ExtractedDataTypes`
   - `getFullResource` and `getFullResourceBySlug` methods now also retrieve generated questions and answers
+- Support the `conversation` RAG strategy
 
 ### Bug fix